Complete adding Swagger and Exception Handler

+++ src/main/java/com/example/UltiOauth/Exception/UserNotFoundException.java
@@ -4,9 +4,10 @@ import java.util.NoSuchElementException;
 
 public class UserNotFoundException extends NoSuchElementException {
 
-  public UserNotFoundException(int userId) {
+  public UserNotFoundException(long userId) {
     super(String.format("User %d not found", userId));
   }
+
 
   public UserNotFoundException(String message) {
     super(message);
